+++ src/main/java/tornadofx/FX.kt
@@ -122,7 +122,7 @@ class FX {
                 if (noArgsConstructor) {
                     replacement = find(obsolete.javaClass.kotlin)
                 } else {
-                    log.warning("Unable to reload $obsolete because it's missing a no args constructor")
+                    log.warning("Unable to reload $obsolete (${obsolete.title}) because it's missing a no args constructor")
                     return
                 }
             }
@@ -137,9 +137,13 @@ class FX {
                     add(index, replacement.root)
                 }
             } else {
-                val scene = obsolete.properties["tornadofx.scene"] as Scene
-                replacement.properties["tornadofx.scene"] = scene
-                scene.root = replacement.root
+                if (obsolete.properties.containsKey("tornadofx.scene")) {
+                    val scene = obsolete.properties["tornadofx.scene"] as Scene
+                    replacement.properties["tornadofx.scene"] = scene
+                    scene.root = replacement.root
+                } else {
+                    log.warning("Unable to reload $obsolete (${obsolete.title}) because it has no parent and no scene attached")
+                }
             }
         }
     }