+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-[Unreleased]
+[1.4.5] - 2016-04-29
 
 ### Changed